@@ -88,50 +87,105 @@
     "store2": "^2.14.4",
     "string-strip-html": "8.3.0",
     "uuid": "^9.0.1",
-    "vue": "^2.6.12",
+    "vue": "~2.7.16",
     "vue-croppa": "^1.3.8",
     "vue-custom-element": "https://github.com/learningequality/vue-custom-element.git#master",
     "vue-intl": "^3.0.0",
     "vue-router": "3.6.5",
     "vuetify": "^1.5.24",
     "vuex": "^3.0.1",
-    "workbox-precaching": "^7.1.0",
-    "workbox-window": "^7.0.0"
+    "workbox-core": "^7.3.0",
+    "workbox-precaching": "^7.3.0",
+    "workbox-routing": "^7.3.0",
+    "workbox-strategies": "^7.3.0",
+    "workbox-window": "^7.3.0"
   },
   "devDependencies": {
-    "@vue/test-utils": "1.0.0-beta.29",
+    "@babel/core": "^7.26.8",
+    "@babel/plugin-syntax-import-assertions": "^7.26.0",
+    "@babel/plugin-transform-runtime": "^7.26.8",
+    "@babel/preset-env": "^7.26.8",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/user-event": "^14.6.1",
+    "@testing-library/vue": "^5",
+    "@vue/test-utils": "^1.3.6",
     "aphrodite": "https://github.com/learningequality/aphrodite.git",
     "autoprefixer": "^10.4.19",
-    "babel-jest": "^26.0.1",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.2.1",
     "circular-dependency-plugin": "^5.2.0",
+    "css-loader": "7.1.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-jest-dom": "^5.5.0",
+    "eslint-plugin-kolibri": "^0.18.0",
+    "eslint-plugin-vue": "^9.32.0",
     "eslint-import-resolver-webpack": "0.13.8",
-    "fake-indexeddb": "^4.0.0",
+    "fake-indexeddb": "^5.0.2",
     "file-loader": "^6.2.0",
     "flush-promises": "^1.0.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.0.1",
+    "jest": "^29.7.0",
+    "jest-cli": "^29.7.0",
     "jest-each": "^29.7.0",
-    "jest-environment-jsdom-sixteen": "^1.0.3",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-serializer-vue": "^3.1.0",
-    "kolibri-tools": "0.16.0-dev.3",
+    "kolibri-format": "1.0.0",
+    "kolibri-tools": "0.18.0",
+    "mini-css-extract-plugin": "^2.8.1",
+    "node-sass": "9.0.0",
     "npm-run-all": "^4.1.3",
+    "postcss-html": "^1.8.0",
+    "postcss-less": "^6.0.0",
+    "postcss-sass": "^0.5.0",
+    "postcss-scss": "^4.0.9",
+    "postcss-loader": "^8.1.1",
+    "sass-loader": "16.0.4",
+    "stylelint": "^15.11.0",
+    "stylelint-config-html": "^1.1.0",
+    "stylelint-config-prettier": "9.0.5",
+    "stylelint-config-recess-order": "4.6.0",
+    "stylelint-config-recommended-scss": "14.1.0",
+    "stylelint-config-sass-guidelines": "11.1.0",
+    "stylelint-config-standard": "34.0.0",
+    "stylelint-csstree-validator": "3.0.0",
+    "stylelint-scss": "5.3.2",
+    "style-loader": "^4.0.0",
     "stylus": "^0.63.0",
-    "stylus-loader": "^7.1.3",
+    "stylus-loader": "^8.1.0",
+    "vue-jest": "^3.0.7",
+    "vue-loader": "15.11.1",
+    "vue-style-loader": "^4.1.3",
+    "vue-template-compiler": "2.7.16",
     "web-streams-polyfill": "^4.0.0",
-    "workbox-webpack-plugin": "^7.0.0"
+    "webpack": "^5.97.1",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.0",
+    "webpack-merge": "^6.0.1",
+    "workbox-webpack-plugin": "^7.3.0"
   },
   "false": {},
-  "peerDependencies": {},
   "engines": {
-    "node": "16.x",
-    "npm": ">= 8"
+    "node": "18.x"
   },
   "browserslist": [
     "> 1%",
     "Firefox ESR"
   ],
+  "packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af",
   "volta": {
-    "node": "16.20.2",
-    "yarn": "1.22.22"
+    "node": "18.20.7",
+    "pnpm": "10.4.1"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "core-js",
+      "deasync",
+      "es5-ext",
+      "node-sass"
+    ]
   }
 }
