@@ -1,3 +1,3 @@
 // As recommended by https://jestjs.io/docs/en/webpack#handling-static-assets
 
-module.exports = 'test-file-stub';
+export default 'test-file-stub';
